moved tests around folders
